# docs/swagger.yaml
definitions:
  model.DataHistoryJson:
    properties:
      ID:
        type: integer
      createTime:
        type: integer
      engineeringID:
        type: integer
      modelID:
        type: integer
      wellID:
        type: integer
    type: object
  model.Device:
    properties:
      ID:
        type: integer
      engineeringID:
        type: integer
      nameOfDevice:
        type: string
      numberOfDevice:
        type: integer
      typeOfDevice:
        type: string
    type: object
  model.EngineeringJson:
    properties:
      ID:
        type: integer
      address:
        type: string
      beginTime:
        type: integer
      constructionUnit:
        type: string
      devices:
        items:
          $ref: '#/definitions/model.Device'
        type: array
      estimatedCompletionTime:
        type: integer
      head:
        type: string
      name:
        type: string
      numberOfConstructors:
        type: integer
      progress:
        type: number
      state:
        type: integer
      wellName:
        type: string
    type: object
  model.JsonModel:
    properties:
      ID:
        type: integer
      createTime:
        type: integer
      name:
        type: string
      params:
        $ref: '#/definitions/model.ParamsJson'
      useCnt:
        type: integer
    type: object
...
  model.ParamsExtra:
    properties:
      activation:
        type: string
      anomaly_ratio:
        type: string
      batch_size:
        type: string
      checkpoints:
        type: string
      d_ff:
        type: string
      d_model:
        type: string
      devices:
        type: string
      distil:
        type: string
      dropout:
        type: string
      embed:
        type: string
      freq:
        type: string
      gpu:
        type: string
      id:
        type: integer
      learning_rate:
        type: string
      loss:
        type: string
      lradj:
        type: string
      mask_rate:
        type: string
      moving_avg:
        type: string
      n_heads:
        type: string
      num_kernels:
        type: string
      num_workers:
        type: string
      output_attention:
        type: string
      p_hidden_dims:
        type: string
      p_hidden_layers:
        type: string
      patience:
        type: string
      seasonal_patterns:
        type: string
      top_k:
        type: string
      train_epochs:
        type: string
      use_amp:
        type: string
      use_gpu:
        type: string
      use_multi_gpu:
        type: string
      w_lin:
        type: string
    type: object
  model.ParamsJson:
    properties:
      pe:
        $ref: '#/definitions/model.ParamsExtra'
      pu:
        $ref: '#/definitions/model.ParamsUsual'
      useExtra:
        type: boolean
    type: object
  model.ParamsUsual:
    properties:
      c_out:
        type: string
      d_layers:
        type: string
      data:
        type: string
      data_path:
        type: string
      data_test_path:
        type: string
      data_train_path:
        type: string
      data_vali_path:
        type: string
      dec_in:
        type: string
      des:
        type: string
      e_layers:
        type: string
      enc_in:
        type: string
      factor:
        type: string
      features:
        type: string
      id:
        type: integer
      is_training:
        type: string
      itr:
        type: string
      label_len:
        type: string
      model:
        type: string
      model_id:
        type: string
      optim:
        type: string
      pred_len:
        type: string
      root_path:
        type: string
      scale:
        type: string
      seq_len:
        type: string
      target:
        type: string
      task_name:
        type: string
      use_kafka:
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/data/dc:
    post:
      parameters:
      - description: 历史数据id
        in: formData
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 关闭查看历史数据进程
      tags:
      - 历史数据模块
  /api/v1/data/delete:
    post:
      parameters:
      - description: 使用模型时产生的历史数据id
        in: formData
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除历史数据
      tags:
      - 历史数据模块
  /api/v1/data/do:
    post:
      parameters:
      - description: 历史数据id
        in: formData
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 打开查看历史数据进程
      tags:
      - 历史数据模块
  /api/v1/data/new:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 真实数据文件
        in: formData
        name: fileTrue
        required: true
        type: file
      - description: 预测数据文件
        in: formData
        name: filePredict
        required: true
        type: file
      - description: 使用模型时产生的历史数据id
        in: body
        name: history
        required: true
        schema:
          $ref: '#/definitions/model.DataHistoryJson'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 新建历史数据
      tags:
      - 历史数据模块
  /api/v1/data/range:
    get:
      parameters:
      - description: 历史数据id
        in: query
        name: id
        required: true
        type: integer
      - description: 起始时间
        in: query
        name: from
        required: true
        type: integer
      - description: 截止时间
        in: query
        name: to
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取范围历史数据,0,0可以查看全部数据
      tags:
      - 历史数据模块
...
  /api/v1/en/{id}:
    get:
      parameters:
      - description: 施工ID
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: 查看施工详细信息
      tags:
      - Engineering
...
  /api/v1/en/delete:
    post:
      parameters:
      - description: 要删除的施工ID
        in: formData
        name: engineeringID
        required: true
        type: string
      responses: {}
      summary: 删除施工
      tags:
      - Engineering
  /api/v1/en/device/add:
    post:
      parameters:
      - description: 施工设备
        in: body
        name: dd
        required: true
        schema:
          items:
            $ref: '#/definitions/model.Device'
          type: array
      responses: {}
      summary: 增加施工设备
      tags:
      - Engineering
  /api/v1/en/device/delete:
    post:
      parameters:
      - description: 要删除的设备ID
        in: body
        name: deviceIDs
        required: true
        schema:
          items:
            type: string
          type: array
      responses: {}
      summary: 删除施工设备
      tags:
      - Engineering
  /api/v1/en/new:
    post:
      parameters:
      - description: 施工信息
        in: body
        name: ed
        required: true
        schema:
          $ref: '#/definitions/model.EngineeringJson'
      responses: {}
      summary: 新增施工信息
      tags:
      - Engineering
  /api/v1/en/re:
    post:
      parameters:
      - description: 施工信息
        in: body
        name: ed
        required: true
        schema:
          $ref: '#/definitions/model.EngineeringJson'
      responses: {}
      summary: 修改施工信息
      tags:
      - Engineering
  /api/v1/file/:
    get:
      consumes:
      - multipart/form-data
      parameters:
      - description: 文件路径
        in: query
        name: filePath
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 查看对应路径下有哪些文件
      tags:
      - 文件模块
...
  /api/v1/file/csv:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: csv文件
        in: formData
        name: file
        required: true
        type: file
      - description: 文件路径
        in: formData
        name: filePath
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 上传数据集文件到对应路径
      tags:
      - 文件模块
  /api/v1/file/df:
    post:
      parameters:
      - description: 文件路径
        in: formData
        name: filePath
        required: true
        type: string
      - description: 文件名称
        in: formData
        name: fileName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除数据文件
      tags:
      - 文件模块
  /api/v1/file/dp:
    post:
      parameters:
      - description: 文件路径
        in: formData
        name: filePath
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除数据文件夹
      tags:
      - 文件模块
  /api/v1/file/new:
    post:
      parameters:
      - description: 文件路径
        in: formData
        name: filePath
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 新建数据集文件夹
      tags:
      - 文件模块
...
  /api/v1/md/copy:
    post:
      parameters:
      - description: 模型ID
        in: formData
        name: modelID
        required: true
        type: string
      - description: 新模型名称
        in: formData
        name: name
        required: true
        type: string
      responses: {}
      summary: 复制模型
      tags:
      - Model
  /api/v1/md/delete:
    post:
      parameters:
      - description: 模型信息
        in: formData
        name: modelID
        required: true
        type: string
      responses: {}
      summary: 删除模型
      tags:
      - Model
  /api/v1/md/new:
    post:
      parameters:
      - description: 模型信息
        in: body
        name: md
        required: true
        schema:
          $ref: '#/definitions/model.JsonModel'
      responses: {}
      summary: 新增模型
      tags:
      - Model
  /api/v1/md/params:
    get:
      parameters:
      - description: 模型ID
        in: query
        name: modelID
        required: true
        type: string
      responses: {}
      summary: 获取模型参数
      tags:
      - Model
  /api/v1/md/revise:
    post:
      parameters:
      - description: 模型ID
        in: formData
        name: modelID
        required: true
        type: string
      - description: 模型新名称
        in: formData
        name: name
        required: true
        type: string
      - description: 模型use-kafka按钮
        in: formData
        name: useKafka
        required: true
        type: integer
      responses: {}
      summary: 修改模型
      tags:
      - Model
  /api/v1/md/umf:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: pth文件
        in: formData
        name: file
        required: true
        type: file
      - description: 模型id
        in: formData
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 上传checkPoint.pth文件到对应模型下
      tags:
      - Model
  /api/v1/md/use:
    post:
      parameters:
      - description: 模型ID
        in: formData
        name: modelID
        required: true
        type: string
      responses: {}
      summary: 使用模型
      tags:
      - Model
...
swagger: "2.0"